refactor: fix dependency inversion (hooksStore -> swap)
