+++ apps/cowswap-frontend/src/pages/Hooks/index.tsx
@@ -1,7 +1,14 @@
 import React from 'react'
 
-import { SwapPage } from 'pages/Swap'
+import { PostHookButton, PreHookButton } from 'modules/hooksStore'
+import { SwapUpdaters } from 'modules/swap'
+import { SwapWidget } from 'modules/swap'
 
 export function HooksPage() {
-  return <SwapPage hooksEnabled={true} />
+  return (
+    <>
+      <SwapUpdaters />
+      <SwapWidget topContent={<PreHookButton />} bottomContent={<PostHookButton />} />
+    </>
+  )
 }
