+++ apps/cowswap-frontend/src/pages/Swap/index.tsx
@@ -5,21 +5,14 @@ import { useWalletInfo } from '@cowprotocol/wallet'
 
 import { Navigate, useLocation, useParams } from 'react-router-dom'
 
-import { AppDataUpdater } from 'modules/appData'
-import { SwapWidget, SwapDerivedStateUpdater, SwapAmountsFromUrlUpdater, SmartSlippageUpdater } from 'modules/swap'
-import { useSwapSlippage } from 'modules/swap/hooks/useSwapSlippage'
+import { SwapUpdaters, SwapWidget } from 'modules/swap'
 import { getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
 import { parameterizeTradeRoute } from 'modules/trade/utils/parameterizeTradeRoute'
 
 import { Routes } from 'common/constants/routes'
 
-export interface SwapPageProps {
-  hooksEnabled: boolean
-}
-
-export function SwapPage({ hooksEnabled }: SwapPageProps) {
+export function SwapPage() {
   const params = useParams()
-  const slippage = useSwapSlippage()
 
   if (!params.chainId) {
     return <SwapPageRedirect />
@@ -27,11 +20,8 @@ export function SwapPage({ hooksEnabled }: SwapPageProps) {
 
   return (
     <>
-      <AppDataUpdater orderClass="market" slippage={slippage} />
-      <SwapDerivedStateUpdater />
-      <SwapAmountsFromUrlUpdater />
-      <SmartSlippageUpdater />
-      <SwapWidget hooksEnabled={hooksEnabled} />
+      <SwapUpdaters />
+      <SwapWidget />
     </>
   )
 }
@@ -60,7 +50,7 @@ function SwapPageRedirect() {
       outputCurrencyAmount: undefined,
       orderKind: undefined,
     },
-    Routes.SWAP
+    Routes.SWAP,
   )
 
   return <Navigate to={{ ...location, pathname, search: searchParams.toString() }} />
